Updated to valid SPDX document
Notes:
The image is downloaded using Docker:
docker pull
us.gcr.io/k8s-artifacts-prod/autoscaling/cluster-autoscaler:v1.17.2
Sometimes, these images are not browsable. This one, however,
can be found in this general location:
https://console.cloud.google.com/gcr/images/k8s-artifacts-prod
Currently, the domain itself is in flux, so this link may
not be reachable in the future. The current plan is to
move us.gcr.io/k8s-artifacts-prod to just k8s.gcr.io.

Signed-off-by: Nisha K <nishak@vmware.com>